move data layer deps to CourseDataModule
